Added validation and added data.json Should work

+++ S308.FinalProject.Group4/FitnessClub/Members.cs
@@ -11,8 +11,8 @@ namespace FitnessClub
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string CreditCardType { get; set; }
-        public int CreditCardNumber { get; set; }
-        public int PhoneNo { get; set; }
+        public long CreditCardNumber { get; set; }
+        public double PhoneNo { get; set; }
         public string Email { get; set; }
         public string Gender { get; set; }
         public int Age { get; set; }
